% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_calibration.R
\name{add_cal_parm}
\alias{add_cal_parm}
\title{Add calibration parameter}
\usage{
add_cal_parm(
  calibration_df = NULL,
  param,
  change_type,
  val,
  conds = NULL,
  layers = NULL,
  dates = NULL
)
}
\arguments{
\item{calibration_df}{\code{NULL} or tibble from previous \code{add_cal_parm}}

\item{param}{Parameter name}

\item{change_type}{One of absval, abschng, pctchng,}

\item{val}{Value of change}

\item{conds}{Not implemented}

\item{layers}{Numeric vector length two, first to last soil layer}

\item{dates}{Date vector of length two, first to last date for param to apply}
}
\description{
Add calibration parameter
}
\details{
This function is used to build a dataframe for calibration, specifically
to write the calibration file calibration.cal using write_calibration_cal().
The function is pipe-friendly for adding multiple parameters.
}
\examples{
calibration_df <- add_cal_parm(param = "cn2", change_type = "pctchng", val = 10)
calibration_df2 <- calibration_df \%>\%
  add_cal_parm(param = "cn3", change_type = "pctchng", val = 20)
}
