% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_calibration.R
\name{get_NSE}
\alias{get_NSE}
\title{Get Nash-Sutcliffe Efficiency}
\usage{
get_NSE(x_obs, x_sim)
}
\arguments{
\item{x_obs}{Vector of observations}

\item{x_sim}{Vector of simulated data}
}
\value{
Returns the NSE between \code{x_obs} and \code{x_sim}
}
\description{
Get Nash-Sutcliffe Efficiency
}
\examples{
set.seed(100)
x_obs <- 1:10 + rnorm(10)
x_sim <- 1:10
get_NSE(x_obs, x_sim)
get_NSE(x_obs, x_sim - 5)
}
