% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat.R
\name{read_swat_data}
\alias{read_swat_data}
\title{Read SWAT+ output data}
\usage{
read_swat_data(
  filename,
  path,
  vars = "all",
  swat_units = "all",
  date_range = "all"
)
}
\arguments{
\item{filename}{Name of SWAT+ output file}

\item{path}{Path to SWAT+ output files}

\item{vars}{Variables to read from the file. See details}

\item{swat_units}{Vector of spatial units to select from the "unit" column}

\item{date_range}{Date vector (length 2) specifying the date range to read}
}
\value{
The function returns a \code{data.table} object, which is useful for large
datasets. If desired, it can easily be converted to a \code{data.frame} using
\code{as.data.frame}.
}
\description{
Read SWAT+ output data
}
\details{
This function is used to read SWAT+ output data into a data.frame format.
Data is read from the \code{filename} in the \code{path} directory. The
\code{vars} parameter specifies the columns to read and the \code{swat_units}
specifies the units to select.

If \code{vars} is set to \code{"names"}, the function returns a vector of
variable names contained in the file.
}
