% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swat.R
\name{run_swat}
\alias{run_swat}
\title{Run SWAT executable file}
\usage{
run_swat(scenario_path)
}
\arguments{
\item{scenario_path}{Path containing TxtInOut directory with SWAT executable}
}
\description{
Run SWAT executable file
}
\details{
Runs the first executable file in \code{scenario_path}
}
