% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_calibration.R
\name{write_calibration_cal}
\alias{write_calibration_cal}
\title{Write calibration file}
\usage{
write_calibration_cal(swatTxtInOut_path, calibration_df)
}
\arguments{
\item{swatTxtInOut_path}{Path to SWAT scenario directory}

\item{calibration_df}{data.frame with calibration data (see details)}
}
\description{
Write calibration file
}
\details{
This function writes the calibration file calibration.cal in the directory
\code{swatTxtInOut_path}.

The input parameter \code{calibration_df} should contain columns created by
\code{add_cal_parm}, i.e.:
\itemize{
\item NAME
\item CHG_TYPE - one of "absval", "abschg", "pctchg"
\item VAL
\item CONDS
\item LYR1
\item LYR2
\item YEAR1
\item YEAR2
\item DAY1
\item DAY2
\item OBJ_TOT
}
}
\examples{
calibration_df <-
  add_cal_parm(param = "cn2", change_type = "pctchng", val = 10) \%>\%
  add_cal_parm(param = "cn3", change_type = "pctchng", val = 20)
write_calibration_cal(tempdir(), calibration_df)
readLines(file.path(tempdir(), "calibration.cal"))
}
